+++ providers/structured/provider.go
@@ -5,118 +5,134 @@ package structured
 
 import (
 	"context"
+	"sync"
 
 	"github.com/myhelix/contextlogger/log"
 	"github.com/myhelix/contextlogger/providers"
 	"github.com/myhelix/contextlogger/providers/chaining"
 )
 
-type RawLogCallArgs struct {
+type LogCallArgs struct {
 	ContextFields log.Fields
 	Report        bool
 	Args          []interface{}
+	Level         providers.LogLevel
 }
 
 type RecordCallArgs struct {
 	ContextFields log.Fields
 	Metrics       log.Metrics
-}
-
-type RecordEventCallArgs struct {
-	ContextFields log.Fields
 	EventName     string
-	Metrics       log.Metrics
 }
 
 type StructuredOutputLogProvider struct {
 	providers.LogProvider
 
-	rawLogCalls      map[providers.RawLogCallType][]RawLogCallArgs
-	recordCalls      []RecordCallArgs
-	recordEventCalls []RecordEventCallArgs
+	logCalls    []*LogCallArgs
+	logMutex    sync.RWMutex
+	recordCalls []*RecordCallArgs
+	recordMutex sync.RWMutex
 }
 
-func (p StructuredOutputLogProvider) GetRawLogCalls() map[providers.RawLogCallType][]RawLogCallArgs {
-	return p.rawLogCalls
+// Return list of log calls, filtered to only selected levels (if any present)
+func (p StructuredOutputLogProvider) LogCalls(levels ...providers.LogLevel) (result []*LogCallArgs) {
+	p.logMutex.RLock()
+	defer p.logMutex.RUnlock()
+
+	allLevels := len(levels) == 0
+	for _, call := range p.logCalls {
+		if allLevels {
+			result = append(result, call)
+		} else {
+			for _, level := range levels {
+				if call.Level == level {
+					result = append(result, call)
+					break
+				}
+			}
+		}
+	}
+	return
 }
 
-func (p StructuredOutputLogProvider) GetRecordCalls() []RecordCallArgs {
-	return p.recordCalls
-}
+func (p StructuredOutputLogProvider) RecordCalls() []*RecordCallArgs {
+	p.recordMutex.RLock()
+	defer p.recordMutex.RUnlock()
 
-func (p StructuredOutputLogProvider) GetRecordEventCalls() []RecordEventCallArgs {
-	return p.recordEventCalls
-}
+	result := make([]*RecordCallArgs, len(p.recordCalls))
 
-// NewStructuredOutputLogProvider returns a LogProvider which records all calls made to it.
-// Deprecated: Clients should call LogProvider instead
-func NewStructuredOutputLogProvider() *StructuredOutputLogProvider {
-	return LogProvider(nil)
+	for i, call := range p.recordCalls {
+		result[i] = call
+	}
+	return result
 }
 
 func LogProvider(nextProvider providers.LogProvider) *StructuredOutputLogProvider {
-	rawLogCalls := map[providers.RawLogCallType][]RawLogCallArgs{}
-	for _, callType := range providers.RawLogCallTypes() {
-		rawLogCalls[callType] = []RawLogCallArgs{}
-	}
-
 	return &StructuredOutputLogProvider{
-		LogProvider:      chaining.LogProvider(nextProvider),
-		rawLogCalls:      rawLogCalls,
-		recordCalls:      []RecordCallArgs{},
-		recordEventCalls: []RecordEventCallArgs{},
+		LogProvider: chaining.LogProvider(nextProvider),
+		logCalls:    []*LogCallArgs{},
+		recordCalls: []*RecordCallArgs{},
 	}
 }
 
-func (p *StructuredOutputLogProvider) saveRawCallArgs(
-	callType providers.RawLogCallType,
+func (p *StructuredOutputLogProvider) saveLogCallArgs(
+	level providers.LogLevel,
 	ctx context.Context,
 	report bool,
 	args ...interface{},
 ) {
-	callArgs := RawLogCallArgs{
+	callArgs := LogCallArgs{
 		ContextFields: log.FieldsFromContext(ctx),
 		Report:        report,
 		Args:          args,
+		Level:         level,
 	}
-	p.rawLogCalls[callType] = append(p.rawLogCalls[callType], callArgs)
+	p.logMutex.Lock()
+	defer p.logMutex.Unlock()
+	p.logCalls = append(p.logCalls, &callArgs)
 }
 
 func (p *StructuredOutputLogProvider) Error(ctx context.Context, report bool, args ...interface{}) {
-	p.saveRawCallArgs(providers.Error, ctx, report, args...)
+	p.saveLogCallArgs(providers.Error, ctx, report, args...)
 	p.LogProvider.Error(ctx, report, args...)
 }
 
 func (p *StructuredOutputLogProvider) Warn(ctx context.Context, report bool, args ...interface{}) {
-	p.saveRawCallArgs(providers.Warn, ctx, report, args...)
+	p.saveLogCallArgs(providers.Warn, ctx, report, args...)
 	p.LogProvider.Warn(ctx, report, args...)
 }
 
 func (p *StructuredOutputLogProvider) Info(ctx context.Context, report bool, args ...interface{}) {
-	p.saveRawCallArgs(providers.Info, ctx, report, args...)
+	p.saveLogCallArgs(providers.Info, ctx, report, args...)
 	p.LogProvider.Info(ctx, report, args...)
 }
 
 func (p *StructuredOutputLogProvider) Debug(ctx context.Context, report bool, args ...interface{}) {
-	p.saveRawCallArgs(providers.Debug, ctx, report, args...)
+	p.saveLogCallArgs(providers.Debug, ctx, report, args...)
 	p.LogProvider.Debug(ctx, report, args...)
 }
 
 func (p *StructuredOutputLogProvider) Record(ctx context.Context, metrics map[string]interface{}) {
+	p.LogProvider.Record(ctx, metrics)
+
 	callArgs := RecordCallArgs{
 		ContextFields: log.FieldsFromContext(ctx),
 		Metrics:       metrics,
 	}
-	p.recordCalls = append(p.recordCalls, callArgs)
-	p.LogProvider.Record(ctx, metrics)
+	p.recordMutex.Lock()
+	defer p.recordMutex.Unlock()
+	p.recordCalls = append(p.recordCalls, &callArgs)
 }
 
 func (p *StructuredOutputLogProvider) RecordEvent(ctx context.Context, eventName string, metrics map[string]interface{}) {
-	callArgs := RecordEventCallArgs{
+	p.LogProvider.RecordEvent(ctx, eventName, metrics)
+
+	callArgs := RecordCallArgs{
 		ContextFields: log.FieldsFromContext(ctx),
 		EventName:     eventName,
 		Metrics:       metrics,
 	}
-	p.recordEventCalls = append(p.recordEventCalls, callArgs)
-	p.LogProvider.RecordEvent(ctx, eventName, metrics)
+	p.recordMutex.Lock()
+	defer p.recordMutex.Unlock()
+	p.recordCalls = append(p.recordCalls, &callArgs)
 }