+++ providers/structured/provider_test.go
@@ -28,8 +28,14 @@ var _ = Describe("bufferedLogProvider", func() {
 		contextLogger log.ContextLogger
 	)
 
+	var verifyEmptyLogCalls = func(logLevels []providers.LogLevel) {
+		for _, level := range logLevels {
+			Ω(provider.LogCalls(level)).Should(BeEmpty())
+		}
+	}
+
 	BeforeEach(func() {
-		provider = NewStructuredOutputLogProvider()
+		provider = LogProvider(nil)
 		contextLogger = log.WithFields(fields)
 	})
 
@@ -39,18 +45,23 @@ var _ = Describe("bufferedLogProvider", func() {
 		provider.Error(contextLogger, false, "Message 2", "Some additional details about message 2")
 
 		// Verify
-		logCalls := provider.GetRawLogCalls()[providers.Error]
-		Ω(len(logCalls)).Should(Equal(2))
-		Ω(logCalls[0]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        true,
-			Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+		errorCalls := provider.LogCalls(providers.Error)
+		Ω(errorCalls).Should(Equal([]*LogCallArgs{
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        true,
+				Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+				Level:         providers.Error,
+			},
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        false,
+				Args:          []interface{}{"Message 2", "Some additional details about message 2"},
+				Level:         providers.Error,
+			},
 		}))
-		Ω(logCalls[1]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        false,
-			Args:          []interface{}{"Message 2", "Some additional details about message 2"},
-		}))
+
+		verifyEmptyLogCalls([]providers.LogLevel{providers.Info, providers.Warn, providers.Debug})
 	})
 
 	It("Should log multiple Warn calls", func() {
@@ -59,18 +70,23 @@ var _ = Describe("bufferedLogProvider", func() {
 		provider.Warn(contextLogger, false, "Message 2", "Some additional details about message 2")
 
 		// Verify
-		logCalls := provider.GetRawLogCalls()[providers.Warn]
-		Ω(len(logCalls)).Should(Equal(2))
-		Ω(logCalls[0]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        true,
-			Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+		warnCalls := provider.LogCalls(providers.Warn)
+		Ω(warnCalls).Should(Equal([]*LogCallArgs{
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        true,
+				Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+				Level:         providers.Warn,
+			},
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        false,
+				Args:          []interface{}{"Message 2", "Some additional details about message 2"},
+				Level:         providers.Warn,
+			},
 		}))
-		Ω(logCalls[1]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        false,
-			Args:          []interface{}{"Message 2", "Some additional details about message 2"},
-		}))
+
+		verifyEmptyLogCalls([]providers.LogLevel{providers.Info, providers.Error, providers.Debug})
 	})
 
 	It("Should log multiple Info calls", func() {
@@ -79,18 +95,23 @@ var _ = Describe("bufferedLogProvider", func() {
 		provider.Info(contextLogger, false, "Message 2", "Some additional details about message 2")
 
 		// Verify
-		logCalls := provider.GetRawLogCalls()[providers.Info]
-		Ω(len(logCalls)).Should(Equal(2))
-		Ω(logCalls[0]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        true,
-			Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+		infoCalls := provider.LogCalls(providers.Info)
+		Ω(infoCalls).Should(Equal([]*LogCallArgs{
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        true,
+				Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+				Level:         providers.Info,
+			},
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        false,
+				Args:          []interface{}{"Message 2", "Some additional details about message 2"},
+				Level:         providers.Info,
+			},
 		}))
-		Ω(logCalls[1]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        false,
-			Args:          []interface{}{"Message 2", "Some additional details about message 2"},
-		}))
+
+		verifyEmptyLogCalls([]providers.LogLevel{providers.Error, providers.Warn, providers.Debug})
 	})
 
 	It("Should log multiple Debug calls", func() {
@@ -99,18 +120,23 @@ var _ = Describe("bufferedLogProvider", func() {
 		provider.Debug(contextLogger, false, "Message 2", "Some additional details about message 2")
 
 		// Verify
-		logCalls := provider.GetRawLogCalls()[providers.Debug]
-		Ω(len(logCalls)).Should(Equal(2))
-		Ω(logCalls[0]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        true,
-			Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+		debugCalls := provider.LogCalls(providers.Debug)
+		Ω(debugCalls).Should(Equal([]*LogCallArgs{
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        true,
+				Args:          []interface{}{"Message 1", "Some additional details about message 1"},
+				Level:         providers.Debug,
+			},
+			&LogCallArgs{
+				ContextFields: fields,
+				Report:        false,
+				Args:          []interface{}{"Message 2", "Some additional details about message 2"},
+				Level:         providers.Debug,
+			},
 		}))
-		Ω(logCalls[1]).Should(Equal(RawLogCallArgs{
-			ContextFields: fields,
-			Report:        false,
-			Args:          []interface{}{"Message 2", "Some additional details about message 2"},
-		}))
+
+		verifyEmptyLogCalls([]providers.LogLevel{providers.Info, providers.Warn, providers.Error})
 	})
 
 	It("Should log multiple Record calls", func() {
@@ -125,20 +151,21 @@ var _ = Describe("bufferedLogProvider", func() {
 		})
 
 		// Verify
-		recordCalls := provider.GetRecordCalls()
-		Ω(len(recordCalls)).Should(Equal(2))
-		Ω(recordCalls[0]).Should(Equal(RecordCallArgs{
-			ContextFields: fields,
-			Metrics: log.Metrics{
-				"Metric 1": "Value 1",
-				"Metric 2": "Value 2",
+		recordCalls := provider.RecordCalls()
+		Ω(recordCalls).Should(Equal([]*RecordCallArgs{
+			&RecordCallArgs{
+				ContextFields: fields,
+				Metrics: log.Metrics{
+					"Metric 1": "Value 1",
+					"Metric 2": "Value 2",
+				},
 			},
-		}))
-		Ω(recordCalls[1]).Should(Equal(RecordCallArgs{
-			ContextFields: fields,
-			Metrics: log.Metrics{
-				"Metric 3": "Value 3",
-				"Metric 4": "Value 4",
+			&RecordCallArgs{
+				ContextFields: fields,
+				Metrics: log.Metrics{
+					"Metric 3": "Value 3",
+					"Metric 4": "Value 4",
+				},
 			},
 		}))
 	})
@@ -155,22 +182,23 @@ var _ = Describe("bufferedLogProvider", func() {
 		})
 
 		// Verify
-		recordEventCalls := provider.GetRecordEventCalls()
-		Ω(len(recordEventCalls)).Should(Equal(2))
-		Ω(recordEventCalls[0]).Should(Equal(RecordEventCallArgs{
-			ContextFields: fields,
-			EventName:     "Event1",
-			Metrics: log.Metrics{
-				"Metric 1": "Value 1",
-				"Metric 2": "Value 2",
+		recordEventCalls := provider.RecordCalls()
+		Ω(recordEventCalls).Should(Equal([]*RecordCallArgs{
+			&RecordCallArgs{
+				ContextFields: fields,
+				EventName:     "Event1",
+				Metrics: log.Metrics{
+					"Metric 1": "Value 1",
+					"Metric 2": "Value 2",
+				},
 			},
-		}))
-		Ω(recordEventCalls[1]).Should(Equal(RecordEventCallArgs{
-			ContextFields: fields,
-			EventName:     "Event2",
-			Metrics: log.Metrics{
-				"Metric 3": "Value 3",
-				"Metric 4": "Value 4",
+			&RecordCallArgs{
+				ContextFields: fields,
+				EventName:     "Event2",
+				Metrics: log.Metrics{
+					"Metric 3": "Value 3",
+					"Metric 4": "Value 4",
+				},
 			},
 		}))
 	})
@@ -241,12 +269,12 @@ var _ = Describe("bufferedLogProvider", func() {
 		})
 
 		// Verify logs
-		Ω(lp.GetRawLogCalls()).Should(Equal(provider.GetRawLogCalls()))
+		Ω(lp.LogCalls(providers.Error)).Should(Equal(provider.LogCalls(providers.Error)))
+		Ω(lp.LogCalls(providers.Info)).Should(Equal(provider.LogCalls(providers.Info)))
+		Ω(lp.LogCalls(providers.Debug)).Should(Equal(provider.LogCalls(providers.Debug)))
+		Ω(lp.LogCalls(providers.Warn)).Should(Equal(provider.LogCalls(providers.Warn)))
 
-		// Verify metrics
-		Ω(lp.GetRecordCalls()).Should(Equal(provider.GetRecordCalls()))
-
-		// Verify events
-		Ω(lp.GetRecordEventCalls()).Should(Equal(provider.GetRecordEventCalls()))
+		// Verify events and metrics
+		Ω(lp.RecordCalls()).Should(Equal(provider.RecordCalls()))
 	})
 })